+++ kernel/kernel/kernel.c
@@ -6,6 +6,7 @@
 
 #include <lib/multiboot.h>
 
+#include <stdbool.h>
 #include <stdint.h>
 
 #define CHECK_FLAG(flags,bit)   ((flags) & (1 << (bit)))
@@ -21,23 +22,22 @@ void hcf(void) {
 
 multiboot_info_t *mbi;
 
-void kernel_boot(uint64_t magic, uint64_t addr) {
+void kernel_boot(uint32_t magic, uint32_t addr) {
 
-    if (magic != MULTIBOOT_BOOTLOADER_MAGIC)
-        hcf();
+	if (magic != MULTIBOOT_BOOTLOADER_MAGIC)
+		hcf();
 
-	  mbi = (multiboot_info_t *) addr;
-
+	mbi = (multiboot_info_t *) (uintptr_t) addr;
+    
     gdt_install(); // maybe move these calls to boot.S
     idt_install();
     isrs_install();
     pic_initalize();
 }
 
-// extern unsigned char _binary_libs_VGA[];
 void kernel_main(void) {
     terminal_initalize(mbi);
-    // asm ("int $18");
+	
     for(;;)
         asm ("hlt");
